+++ index.html
@@ -52,6 +52,8 @@
                 
     </main>
     
-    <script src="./src/main.js"></script>
+    <script type="module" src="./src/main.js"></script>
+    <script type="module" src="./src/mudels/ui.js"></script>
+    <script type="module" src="./src/mudels/storage.js"></script>
 </body>
 </html>
