+++ src/journal.html
@@ -38,6 +38,8 @@
 
     <main id="cards-container" class="w-2/3 grid grid-cols-3 auto-rows-min gap-2 overflow-y-auto scrollbar" ></main>
     
-    <script src="journal.js"></script>
+    <script type="module" src="journal.js"></script>
+   
+  
 </body>
 </html>
